This PR fixes a bug that Sophie encountered in requesting a JDP file transfer.
When we request metadata for a given set of file IDs from the JDP, the response needs to be
de-duped for some reason. This is easy enough to do, and I certainly don't want to make any
assumptions about what the JDP REST API is or isn't doing for us, so this fix performs the
de-duping as part of the re-ordering process.

Also, it moves the service's version numbers into its own file for simplicity.